+++ src/main/java/com/invadermonky/pickuplimit/config/ConfigHandlerPL.java
@@ -25,6 +25,9 @@ public class ConfigHandlerPL {
         @Config.Comment("Enables limit tooltips on items. Tooltips must be defined in limit groups builders using the 'setLimitTooltip' method.")
         public boolean enableTooltipLimitMessage = true;
 
+        @Config.Comment("Limit group tooltip messages require the player to hold shift to display. If set to false, they will always display.")
+        public boolean tooltipMessageRequiresShift = true;
+
         @Config.RangeInt(min = 1, max = 12000)
         @Config.Comment("The time, in ticks, between each inventory check. The inventory check will scan a player's inventory to ensure\n" +
                 "they have not exceeded any item limits.")